loading list refactor
