--- components/basecomponents/MemoryGame/ImageSuspense.tsx
@@ -1,17 +1,0 @@
-import React, { ImgHTMLAttributes, useCallback } from 'react';
-import { registerImage } from './loadingList';
-
-const ImageSuspense = ({ src, ...rest }: { src: string } & ImgHTMLAttributes<HTMLImageElement>) => {
-  const loadImg = useCallback(
-    (img) => {
-      if (img !== null) {
-        registerImage(src, img);
-      }
-    },
-    [src]
-  );
-
-  return <img ref={loadImg} src={src} {...rest} />;
-};
-
-export default ImageSuspense;
