--- components/basecomponents/MemoryGame/SuspenseKicker.tsx
@@ -1,10 +1,0 @@
-import { useLoading } from './loadingList';
-
-const SuspenseKicker = () => {
-  const { loading, loadingPromise } = useLoading();
-
-  if (loading && loadingPromise instanceof Promise) throw loadingPromise;
-  return null;
-};
-
-export default SuspenseKicker;
